libfs: move billy implementation into libfs
Suggested by songgao.

Issue: #1108
