+++ appveyor.yml
@@ -64,6 +64,7 @@ build_script:
   - echo github.com/keybase/kbfs/kbfshash >> testlist.txt
   - echo github.com/keybase/kbfs/kbfssync >> testlist.txt
   - echo github.com/keybase/kbfs/tlf >> testlist.txt
+  - echo github.com/keybase/kbfs/libfs >> testlist.txt
   - echo github.com/keybase/kbfs/libkbfs >> testlist.txt
   - echo github.com/keybase/kbfs/libdokan >> testlist.txt
   - echo github.com/keybase/kbfs/kbfsgit >> testlist.txt
